Extract ConfigDatabaseRemoveCommand
