+++ modules/cli/src/main/kotlin/org/coner/trailer/cli/command/ConfigDatabaseCommand.kt
@@ -1,11 +1,9 @@
 package org.coner.trailer.cli.command
 
-import com.github.ajalt.clikt.core.Abort
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
-import com.github.ajalt.clikt.parameters.types.choice
 import com.github.ajalt.clikt.parameters.types.file
 import org.coner.trailer.cli.io.ConfigurationService
 import org.coner.trailer.cli.io.DatabaseConfiguration
@@ -28,7 +26,6 @@ class ConfigDatabaseCommand : CliktCommand(
 """.trimIndent() }
     }
 }
-
 
 class ConfigDatabaseListCommand(
         private val config: ConfigurationService
@@ -78,28 +75,6 @@ class ConfigDatabaseSetCommand(
         )
         config.configureDatabase(dbConfig)
         echo(render(dbConfig))
-    }
-}
-
-class ConfigDatabaseRemoveCommand(
-        private val config: ConfigurationService
-) : CliktCommand(
-        name = "remove",
-        help = """
-                Remove a database from the CLI app's config file.
-                
-                This is a non-destructive operation -- you can always add the database again. This will not affect the
-                database itself.
-            """.trimIndent()
-) {
-
-    val dbConfig: DatabaseConfiguration by option(names = *arrayOf("--name"))
-            .choice(config.listDatabasesByName())
-            .required()
-
-    override fun run() {
-        if (dbConfig == config.noDatabase) throw Abort()
-        config.removeDatabase(dbConfig)
     }
 }
